Prevent VAE from refreshing when info panel hovered
When Simple Calendar is running and the clock is active, VAE refreshes constantly causing stuttering and making it impossible to hover the info panel for more than a second. You can't expand the details section. This pauses the refreshing while the panel is hovered, it resumes on mouseout. The cool thing is if you hover the icon you can still see the timer counting down.